working on updating article

+++ raf-news-backend-jax-rs/src/main/java/rs/raf/rafnews/repository/implementation/CategoryRepository.java
@@ -2,20 +2,25 @@ package rs.raf.rafnews.repository.implementation;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import jakarta.enterprise.context.RequestScoped;
+import jakarta.inject.Inject;
 import rs.raf.rafnews.database.RafNewsDatabase;
 import rs.raf.rafnews.dto.CategoryDto;
 import rs.raf.rafnews.dto.ServiceUserDto;
 import rs.raf.rafnews.exception.*;
 import rs.raf.rafnews.model.Category;
+import rs.raf.rafnews.model.ServiceUser;
 import rs.raf.rafnews.repository.specification.ICategoryRepository;
+import rs.raf.rafnews.service.specification.IArticleService;
 
+import java.lang.reflect.Field;
 import java.sql.*;
 import java.util.ArrayList;
 import java.util.List;
 
 @RequestScoped
 public class CategoryRepository implements ICategoryRepository {
-
+    @Inject
+    private IArticleService articleService;
     @Override
     public List<Category> getAllRawCategories() throws JsonProcessingException, GetException, SQLException {
         Connection connection = RafNewsDatabase.getInstance().getConnection();
@@ -147,12 +152,39 @@ public class CategoryRepository implements ICategoryRepository {
 
     @Override
     public Integer updateCategoryById(Integer id, Category category) throws SQLException, JsonProcessingException, UpdateException {
+        try {
+            Category currentCategory = getRawCategoryById(id);
+            if(currentCategory.getId() > 0){
+                Class<?> categoryClass = Category.class;
+                Field[] fields = categoryClass.getDeclaredFields();
+                for (Field field : fields) {
+                    field.setAccessible(true);
+                    Object newValue = field.get(category);
+                    if (newValue != null) {
+                        field.set(currentCategory, newValue);
+                    }
+                }
+                return saveCategory(currentCategory);
+            }
+            else{
+                ExceptionMessage exceptionMessage = new ExceptionMessage("UpdateException", "Failed to update category with id: " + id + " with new category: " + category);
+                throw new UpdateException(exceptionMessage);
+            }
+        }
+        catch (Exception e) {
+            ExceptionMessage exceptionMessage = new ExceptionMessage("UpdateException", "Failed to update category with id: " + id + " with new category: " + category + ". " + e.getMessage());
+            throw new UpdateException(exceptionMessage);
+        }
+    }
+
+    @Override
+    public Integer saveCategory(Category category) throws SQLException, UpdateException, JsonProcessingException {
         Connection connection = RafNewsDatabase.getInstance().getConnection();
         String query = "UPDATE category SET category_name = ?, description = ? WHERE id = ?";
         try (PreparedStatement preparedStatement = connection.prepareStatement(query)){
             preparedStatement.setString(1, category.getCategory_name());
             preparedStatement.setString(2, category.getDescription());
-            preparedStatement.setInt(3, id);
+            preparedStatement.setInt(3, category.getId());
             int affectedRows = preparedStatement.executeUpdate();
             if(affectedRows >= 0){
                 return affectedRows;
@@ -169,26 +201,30 @@ public class CategoryRepository implements ICategoryRepository {
     }
 
     @Override
-    public Integer deleteCategoryById(Integer id) throws JsonProcessingException, DeleteException, SQLException {
-        Connection connection = RafNewsDatabase.getInstance().getConnection();
-        try{
-            String query = "DELETE FROM category WHERE id = ?";
-            try(PreparedStatement preparedStatement = connection.prepareStatement(query)){
-                preparedStatement.setInt(1, id);
-                int affectedRows = preparedStatement.executeUpdate();
-                if(affectedRows > 0){
-                    return affectedRows;
-                }
-                else{
-                    ExceptionMessage exceptionMessage = new ExceptionMessage("DeleteException", "Failed to delete category. Id " + id + " not found.");
-                    throw new DeleteException(exceptionMessage);
+    public Integer deleteCategoryById(Integer id) throws JsonProcessingException, DeleteException, SQLException, GetException {
+        if(articleService.getRawArticleByCategoryId(id).getId() < 0){
+            Connection connection = RafNewsDatabase.getInstance().getConnection();
+            try{
+                String query = "DELETE FROM category WHERE id = ?";
+                try(PreparedStatement preparedStatement = connection.prepareStatement(query)){
+                    preparedStatement.setInt(1, id);
+                    int affectedRows = preparedStatement.executeUpdate();
+                    if(affectedRows > 0){
+                        return affectedRows;
+                    }
+                    else{
+                        ExceptionMessage exceptionMessage = new ExceptionMessage("DeleteException", "Failed to delete category. Id " + id + " not found.");
+                        throw new DeleteException(exceptionMessage);
+                    }
                 }
             }
+            catch (SQLException e){
+                ExceptionMessage exceptionMessage = new ExceptionMessage("DeleteException", e.getMessage());
+                throw new DeleteException(exceptionMessage);
+            }
         }
-        catch (SQLException e){
-            ExceptionMessage exceptionMessage = new ExceptionMessage("DeleteException", e.getMessage());
-            throw new DeleteException(exceptionMessage);
-        }
+        ExceptionMessage exceptionMessage = new ExceptionMessage("DeleteException", "Failed to delete category. An article with that category exists.");
+        throw new DeleteException(exceptionMessage);
     }
     private Category extractCategoryFromResultSet(ResultSet resultSet) throws SQLException {
         Integer columnId = resultSet.getInt("id");
@@ -196,4 +232,6 @@ public class CategoryRepository implements ICategoryRepository {
         String columnDescription = resultSet.getString("description");
         return new Category(columnId, columnCategoryName, columnDescription);
     }
+
+
 }
